# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "H&M Fashion Recommendation API",
    "description": "Personalized fashion recommendations via LightGBM LambdaRank.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/customers/sample": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Sample demo customers by segment",
        "operationId": "sample_customers_api_v1_customers_sample_get",
        "parameters": [
          {
            "name": "segment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Segment: cold-start, light, medium, heavy",
              "title": "Segment"
            },
            "description": "Segment: cold-start, light, medium, heavy"
          },
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SampleCustomer"
                  },
                  "title": "Response Sample Customers Api V1 Customers Sample Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{customer_id}": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Customer profile summary",
        "operationId": "customer_profile_api_v1_customers__customer_id__get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{customer_id}/recommendations": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Ranked recommendations for a customer",
        "operationId": "recommendations_api_v1_customers__customer_id__recommendations_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          },
          {
            "name": "k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 12,
              "title": "K"
            }
          },
          {
            "name": "include_reasons",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Reasons"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{customer_id}/baseline": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Heuristic baseline top-k (for side-by-side comparison)",
        "operationId": "baseline_api_v1_customers__customer_id__baseline_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          },
          {
            "name": "k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 12,
              "title": "K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecommendedItemNoReasons"
                  },
                  "title": "Response Baseline Api V1 Customers  Customer Id  Baseline Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{customer_id}/actual-purchases": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Week-104 actual purchases (display only, labelled)",
        "operationId": "actual_purchases_api_v1_customers__customer_id__actual_purchases_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleMeta"
                  },
                  "title": "Response Actual Purchases Api V1 Customers  Customer Id  Actual Purchases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{customer_id}/funnel": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Full candidate funnel: source, score, final rank",
        "operationId": "funnel_api_v1_customers__customer_id__funnel_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FunnelItem"
                  },
                  "title": "Response Funnel Api V1 Customers  Customer Id  Funnel Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{customer_id}/explain/{article_id}": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Full contrast-SHAP breakdown for one (customer, article) pair",
        "operationId": "explain_api_v1_customers__customer_id__explain__article_id__get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          },
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/custom": {
      "post": {
        "tags": [
          "customers"
        ],
        "summary": "Recommendations for a custom history (build-your-own customer)",
        "operationId": "custom_recommendations_api_v1_recommendations_custom_post",
        "parameters": [
          {
            "name": "k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 12,
              "title": "K"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomRecommendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArticleMeta": {
        "properties": {
          "article_idx": {
            "type": "integer",
            "title": "Article Idx"
          },
          "article_id": {
            "type": "string",
            "title": "Article Id"
          },
          "prod_name": {
            "type": "string",
            "title": "Prod Name"
          },
          "product_type": {
            "type": "string",
            "title": "Product Type"
          },
          "colour": {
            "type": "string",
            "title": "Colour"
          },
          "department": {
            "type": "string",
            "title": "Department"
          }
        },
        "type": "object",
        "required": [
          "article_idx",
          "article_id",
          "prod_name",
          "product_type",
          "colour",
          "department"
        ],
        "title": "ArticleMeta"
      },
      "CustomHistoryItem": {
        "properties": {
          "article_id": {
            "type": "string",
            "title": "Article Id"
          },
          "t_dat": {
            "type": "string",
            "format": "date",
            "title": "T Dat"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "sales_channel_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sales Channel Id"
          }
        },
        "type": "object",
        "required": [
          "article_id",
          "t_dat"
        ],
        "title": "CustomHistoryItem"
      },
      "CustomRecommendRequest": {
        "properties": {
          "history": {
            "items": {
              "$ref": "#/components/schemas/CustomHistoryItem"
            },
            "type": "array",
            "minItems": 1,
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "history"
        ],
        "title": "CustomRecommendRequest"
      },
      "CustomerProfile": {
        "properties": {
          "customer_idx": {
            "type": "integer",
            "title": "Customer Idx"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "age_bucket": {
            "type": "string",
            "title": "Age Bucket"
          },
          "n_purchases_all_time": {
            "type": "integer",
            "title": "N Purchases All Time"
          },
          "n_purchases_last_4w": {
            "type": "integer",
            "title": "N Purchases Last 4W"
          },
          "recent_history": {
            "items": {
              "$ref": "#/components/schemas/ArticleMeta"
            },
            "type": "array",
            "title": "Recent History"
          }
        },
        "type": "object",
        "required": [
          "customer_idx",
          "customer_id",
          "age_bucket",
          "n_purchases_all_time",
          "n_purchases_last_4w",
          "recent_history"
        ],
        "title": "CustomerProfile"
      },
      "ExplainResult": {
        "properties": {
          "customer_idx": {
            "type": "integer",
            "title": "Customer Idx"
          },
          "article_idx": {
            "type": "integer",
            "title": "Article Idx"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "expected_value": {
            "type": "number",
            "title": "Expected Value"
          },
          "shap_values": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Shap Values"
          },
          "contrast": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Contrast"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "customer_idx",
          "article_idx",
          "score",
          "expected_value",
          "shap_values",
          "contrast",
          "reasons"
        ],
        "title": "ExplainResult"
      },
      "FunnelItem": {
        "properties": {
          "article_idx": {
            "type": "integer",
            "title": "Article Idx"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "all_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "All Sources"
          },
          "source_rank": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Rank"
          },
          "model_score": {
            "type": "number",
            "title": "Model Score"
          },
          "final_rank": {
            "type": "integer",
            "title": "Final Rank"
          }
        },
        "type": "object",
        "required": [
          "article_idx",
          "source",
          "all_sources",
          "source_rank",
          "model_score",
          "final_rank"
        ],
        "title": "FunnelItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecommendedItemNoReasons": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "article_idx": {
            "type": "integer",
            "title": "Article Idx"
          },
          "article_id": {
            "type": "string",
            "title": "Article Id"
          },
          "prod_name": {
            "type": "string",
            "title": "Prod Name"
          },
          "product_type": {
            "type": "string",
            "title": "Product Type"
          },
          "colour": {
            "type": "string",
            "title": "Colour"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "article_idx",
          "article_id",
          "prod_name",
          "product_type",
          "colour",
          "department",
          "sources"
        ],
        "title": "RecommendedItemNoReasons"
      },
      "SampleCustomer": {
        "properties": {
          "customer_idx": {
            "type": "integer",
            "title": "Customer Idx"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "segment": {
            "type": "string",
            "title": "Segment"
          },
          "age_bucket": {
            "type": "string",
            "title": "Age Bucket"
          }
        },
        "type": "object",
        "required": [
          "customer_idx",
          "customer_id",
          "segment",
          "age_bucket"
        ],
        "title": "SampleCustomer"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}